internal/contour: remove cache and clusterLoadAssignmentCache types
Updates #1091

Remove intermediary types which were only used by EndpointTranslator.
