+++ internal/contour/cluster.go
@@ -18,6 +18,7 @@ import (
 	"sync"
 
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2"
+	"github.com/envoyproxy/go-control-plane/pkg/cache"
 	"github.com/gogo/protobuf/proto"
 	"github.com/heptio/contour/internal/dag"
 	"github.com/heptio/contour/internal/envoy"
@@ -90,7 +91,7 @@ func (c clusterByName) Len() int           { return len(c) }
 func (c clusterByName) Swap(i, j int)      { c[i], c[j] = c[j], c[i] }
 func (c clusterByName) Less(i, j int) bool { return c[i].(*v2.Cluster).Name < c[j].(*v2.Cluster).Name }
 
-func (*ClusterCache) TypeURL() string { return clusterType }
+func (*ClusterCache) TypeURL() string { return cache.ClusterType }
 
 type clusterVisitor struct {
 	clusters map[string]*v2.Cluster
